% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bears.R, R/data.R
\docType{data}
\name{bears}
\alias{bears}
\alias{bears}
\alias{phy_mat}
\alias{tree}
\title{Dataset of Morphological character for extant and extinct bears.}
\format{\describe{
\item{bears}{A fasta dataset frame with 53940 rows and 10 variables}
\item{tree}{A phylo object of a phylogeny}
\item{phy_mat}{Bears dataset formatted as a phyDat object}
}}
\source{
\url{https://github.com/revbayes/revbayes_tutorial/tree/master/RB_TotalEvidenceDating_FBD_Tutorial/data}

\url{https://github.com/revbayes/revbayes_tutorial/tree/master/RB_TotalEvidenceDating_FBD_Tutorial/data}
}
\usage{
bears

data(bears)
}
\description{
A dataset containing 18 taxa and 62 characters for bears.

A dataset containing 18 taxa and 62 characters for bears.
}
\keyword{datasets}
