% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sets.R
\name{generate_sliding}
\alias{generate_sliding}
\title{Generate list of character sets to visualize}
\usage{
generate_sliding(data_set, start_char, stop_char, steps)
}
\arguments{
\item{data_set}{Name of the phylogenetic matrix from which character sets will be sampled}

\item{start_value}{Integer indicating where in the data set to begin generating character sets}

\item{tree}{Integer indicating where in the data set to stop generating character sets}

\item{step}{Integer indicating how amny characters will be in each set}
}
\value{
df
}
\description{
Generate a dataframe of start and stop values. These will determine what character sets from the total phylogenetic matrix will appear in each visualization
}
\examples{
sample_df <- generate_sliding(bears, 1, 10, 2)
}
