% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree_vec.R
\name{generate_tree_vec}
\alias{generate_tree_vec}
\title{Generate a list of parsimony trees, based on subsamples from phylogenetic matrix}
\usage{
generate_tree_vec(data_set, start, stop, tree)
}
\arguments{
\item{data_set}{Phylogenetic matrix}

\item{start}{Which character to begin sampling characters at}

\item{stop}{Which character to end sampling characters at}

\item{starting_tree}{Starting tree for parsimony search}
}
\value{
Vector
}
\description{
Generate a list of phylogenetic trees based on subsamples from phylogenetic matrix, as generated from function generate_sets. Return return a vector of plots showing the tree and the specified data.
}
\examples{
tree <- generate_tree_vec(bears, 1, 2, tree)
}
