% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree_vis.R
\name{generate_tree_vis}
\alias{generate_tree_vis}
\title{Generate a list of parsimony trees, based on subsamples from phylogenetic matrix}
\usage{
generate_tree_vis(data_frame, alignment, tree, data_set, pscore = FALSE,
  lscore = FALSE)
}
\arguments{
\item{alignment}{String specifying path to alignment}

\item{tree}{Starting tree for parsimony analysis. Phylo object.}

\item{data_set}{Phylogenetic matrix to be subsampled}

\item{pscore}{Optional boolean specifying if you would like figures annotated with parsimony score of tree given data. Default FALSE.}

\item{sample_df}{Dataframe specifying how to subsample phylogenetic matrix}

\item{pscore}{Optional boolean specifying if you would like figures annotated with likelihood score of tree given data under the Mk model. Default FALSE.}
}
\value{
vector Vector of ggplot2 objects
}
\description{
Generate a list of phylogenetic trees based on subsamples from phylogenetic matrix, as generated from function generate_sets. Return the input dataframe, with a column of phylogenetic trees.
}
\examples{
output_vector <- generate_tree_vis(data_frame = sample_df, alignment = aln_path, tree = tree, data_set = bears, pscore = TRUE, lscore = TRUE)
}
