% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_dat.R
\name{tree_dat}
\alias{tree_dat}
\title{Generate likelihood and parsimony scores for each tree.}
\usage{
tree_dat(tree, phy_mat, start, stop, pscore = FALSE, lscore = FALSE)
}
\arguments{
\item{tree}{Exported tree from parsimony search}

\item{phy_mat}{Phylogenetic matrix}

\item{start}{Which character to begin sampling characters at}

\item{stop}{Which character to end sampling characters at}

\item{pscore}{Boolean calculate and export parsimony scores for each character set}

\item{lscore}{Boolean calculate and export likelihood scores for each character set}
}
\value{
tree Tree object with p and l scores annotated via $
}
\description{
Generate extra data, including likelihood and parsimony scores for each tree. Export the values as tree annotation to put into the ggplot object.
}
\examples{
data(bears)
tree <- generate_tree_vec(bears, 1, 2, tree)
}
