% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree_vis.R
\name{generate_tree_vis}
\alias{generate_tree_vis}
\title{Combine trees and phylogenetic matrix data as a ggplot object}
\usage{
generate_tree_vis(sample_df, alignment, tree, phy_mat, pscore = FALSE,
  lscore = FALSE)
}
\arguments{
\item{sample_df}{Dataframe specifying how to subsample phylogenetic matrix}

\item{alignment}{String specifying path to alignment}

\item{tree}{Starting tree for parsimony analysis. Phylo object.}

\item{phy_mat}{Phylogenetic matrix to be subsampled}

\item{pscore}{Optional boolean specifying if you would like figures annotated with parsimony score of tree given data. Default FALSE.}

\item{lscore}{Optional boolean specifying if you would like figures annotated with likelihood score of tree given data under the Mk model. Default FALSE.}
}
\value{
vector Vector of ggplot2 objects
}
\description{
This function calls generate_tree_vec() to generate a vector of phylogenetic trees under parsimony. It then plots them alongside the phylogenetic matrix that was used to estimate the tree. Optional decorations can be added displaying parsimony and likelihood scores.
}
\examples{
\dontrun{
output_vector <- generate_tree_vis(sample_df = sample_df,
                                   alignment = aln_path, tree = tree,
                                   phy_mat = bears, pscore = TRUE,
                                   lscore = TRUE)
}
}
